refactor: simplify spend experiences admin and DB patch build
Extract form helpers, list UI, and slide-over form panel from the admin
page. Replace manual updateSpendExperience field checks with a keyed
loop over UpdateSpendExperienceInput fields.
